### 2023-03-20 #5276 @editorialbot
[PRE REVIEW]: CRE: An R package for interpretable discovery and inference of heterogeneous treatment effects
**Submitting author:** @naeemkh (<a href="http://orcid.org/0000-0003-4315-1426">Naeem Khoshnevis</a>)
**Repository:** https://github.com/NSAPH-Software/CRE
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v0.2.1
