### 2023-04-14 #5276 @spholmes
👋 @tedinburgh, this a JOSS submission for an R package that you might find interesting, would you be interested in reviewing this for us? Many Thanks for your time.
