### 2023-05-13 #5276 @spholmes
Hello @gpapadog, would you car to review this submission to JOSS, we have a simple transparent checklist based process, we hope it is interesting and not too onerous, many thanks..
