### 2023-05-13 #5276 @spholmes
👋 @jantonelli111 would you like to review this submission to JOSS, we try to make it a simple transparent  process based on checklist, we hope it is interesting and not too onerous, many thanks..
