### 2023-05-18 #5276 @jantonelli111
Hello, @spholmes, I have a conflict of interest in this case as I am friends with some of the authors and got my PhD in the same group. Can I propose a substitute reviewer, Heejun Shin? He is a 5th year PhD student in my department who works on very related issues. He does not have a Github account, but his email is hshin1@ufl.edu, and I think he would be a great choice for a reviewer.
