### 2023-05-18 #5276 @spholmes
Hi Joseph
Thanks for responding quickly, it might slightly difficult for someone who
has never used Github to do the reviewing process, but for a student it
might be a
good learning experience,
I will get in touch, thanks
Susan

On Thu, May 18, 2023 at 2:38 PM Joseph Antonelli ***@***.***>
wrote:

> Hello, @spholmes <https://github.com/spholmes>, I have a conflict of
> interest in this case as I am friends with some of the authors and got my
> PhD in the same group. Can I propose a substitute reviewer, Heejun Shin? He
> is a 5th year PhD student in my department who works on very related
> issues. He does not have a Github account, but his email is ***@***.***,
> and I think he would be a great choice for a reviewer.
>
