### 2023-06-06 #5276 @spholmes
👋 @carlyls : would you be interested in reviewing this R package presentation for the Journal of Open Source Software. The review process is transparent and based on a checklist and an open discussion with the authors. Many thanks,
