### 2023-06-07 #5276 @Naeemkh
> I am afraid that @Naeemkh, you are creating confusion for the system by suggesting the same reviewers for your two papers, this was the reason for the 24hour lag (not a very long delay).

Dear @spholmes, I appreciate your clarification. It wasn't my intention to create any such confusion, and I value your efforts to streamline the process. I'll bear this in mind for future submissions. Thank you for your guidance.
