### 2023-06-23 #5276 @spholmes
@carlyls : you should be able to start the checklist with the command: @editorialbot generate my checklist
