### 2023-08-16 #5587 @carlyls
@spholmes @Naeemkh
I have finished a preliminary review of the code and the software paper. I think this package is very helpful and well-documented, and I plan to use it in the future! I have some notes below - let me know if I should put these somewhere else. When some of these are addressed, I will feel comfortable checking off more of the boxes in my checklist above. I reference the checklist box with bolded text when relevant.

- **License**: Missing LICENSE file
- **Installation Instructions**: I don’t see a clear list of dependencies in the README
- **Community guidelines**: the code of conduct is good, but can there be something in the README with instructions for contributing or highlighting issues?
- Under the readme -> Notes, can you add citations for the methods listed, like the S-learner etc?
- **Reproducibility**: I can mostly reproduce the plot in the paper using their code but the ATE at the top is different – can you add a seed so that it is reproducible?
- **Reproducibility**: In example 2, "ite_pred <- …" could you put in some code here so that the example can be fully reproduced?
- **Reproducibility**: In example 3 hyper_params, there is not a comma after t_decay = 0.025 in the README
- **Reproducibility**: Also in example 3, when I run plot(cre_results), I get “Visualization not available (0 causal decision rules discovered).” Just want to make sure that is expected.

**Paper**:
- Slightly confusing phrasing at the end of the second sentence in the Summary section
- Statement of need section: why reference causalTree but not the grf package?
- **State of the field**: I think it’s important to refer to other packages like Kunzel’s causalToolbox package (https://github.com/soerenkuenzel/causalToolbox) and Athey’s grf package (https://grf-labs.github.io/grf/REFERENCE.html). It also looks like you use the SuperLearner package? Should that be cited somewhere?
- Would be helpful to mention expected computation time for the cre() function
