### 2023-08-28 #5587 @carlyls
Thank you for the thoughtful responses and for the updates on the new package version!

> On Aug 28, 2023, at 4:30 PM, Naeem Khoshnevis ***@***.***> wrote:
>
>
> Hi @carlyls <https://github.com/carlyls>,
> Thank you for the time you allocated to review our package. All the comments and feedback were extremely valuable and helped us a lot in further improving the package and its documentation.
>
> We provide detailed responses to the comments, addressing each one individually. Additionally, please note that we have been developing a new package version over the past few months. Specifically, we:
>
> consider also internal node nodes for rules generation (to facilitate the discovery of rules with different complexity, i.e. length),
> added the new ‘max_rulesfiltering (while extracting the candidate rules from the trained forest, only consider the topmax_rules` rules most frequent.. This step is useful to pre-filter useless features, speeding up the following steps),
> added (vanilla) Stability Selection (without Error Control) for rules selection,
> added Uncertainty Quantification in estimation by bootstrapping,
> added predict() general function for treatment effect estimation via CRE,
> updated accordingly and simplified several hyper and method parameters.
> For more information about this new release, please refer to https://github.com/NSAPH-Software/CRE/blob/main/NEWS.md <https://github.com/NSAPH-Software/CRE/blob/main/NEWS.md>. All examples have been updated to reflect these changes.
>
> Detailed responses to the questions:
>
> License: Missing LICENSE file
> The package is licensed under GPLv3, and we've included the corresponding license file.
> Installation Instructions: I don’t see a clear list of dependencies in the README
> The complete list of dependencies is available in the DESCRIPTION file. We've also directed users to this information in the README.
> Community guidelines: the code of conduct is good, but can there be something in the README with instructions for contributing or highlighting issues?
> The Community Guidelines can be found on the CRE website <https://nsaph-software.github.io/CRE/articles/Contribution.html>. We've also included a link to this page in the Code of Conduct section of the README.
> Citations: Under the readme -> Notes, can you add citations for the methods listed, like the S-learner etc?
> We added the link to the corresponding paper in the readme for each ITE estimator.
> Reproducibility: (i) I can mostly reproduce the plot in the paper using their code but the ATE at the top is different – can you add a seed so that it is reproducible? (ii) In example 2, "ite_pred <- …" could you put in some code here so that the example can be fully reproduced? (iii) In example 3 hyper_params, there is not a comma after t_decay = 0.025 in the README. (iv) Also in example 3, when I run plot(cre_results), I get “Visualization not available (0 causal decision rules discovered).” Just want to make sure that is expected.
> i. We added a seed.
> ii. We added a customized ITE estimator for this example.
> iii. Fixed.
> iv. If no rules are discovered, that warning is expected. Anyway, in the new examples, cre() should always retrieve more than 1 rule.
> Paper general: Slightly confusing phrasing at the end of the second sentence in the Summary section
> We modified the confusing statement. Also added a few other minor changes according to the last release.
> *Statement of need: why reference causalTree but not the grf package?
> We are not directly comparing our method with grf::causal_forest because we are focusing here only on interpretable methods for HTE.
> State of the field: I think it’s important to refer to other packages like Kunzel’s causalToolbox package (https://github.com/soerenkuenzel/causalToolbox <https://github.com/soerenkuenzel/causalToolbox>) and Athey’s grf package (https://grf-labs.github.io/grf/REFERENCE.html <https://grf-labs.github.io/grf/REFERENCE.html>). It also looks like you use the SuperLearner package? Should that be cited somewhere?
> We added the references to all the R packages we rely on for ITE estimation and Stability Selection. We are not relying on causalToolbox.
> Computation Time: Would be helpful to mention expected computation time for the cre() function
> We added a new section reporting the expected computation time of cre(). Please see Figure 2.
