### 2023-08-30 #5587 @salleuska
@spholmes @Naeemkh

Apologies for the delay. I traveled and moved to a new role over the summer and I had missed some notifications.

I did a preliminary review of the paper and software. I think that the package is interesting, but I think both the paper and package documentation are not immediate to understand.

I'll make another thread for code related issues; @Naeemkh I have been looking at the code in the `main` branch in the past days, but I am assuming that the latest version is in the   `JOSS` one now?

Here my suggestions for the paper and the README.md.

## Software Paper

1. The text in the **Summary** section seems more like a statement of need. I'd suggest revising to better highlight the package features rather than the motivations for the methodology. It is mentioned that the CRE package is a flexible implementation of the  Causal Rule Ensemble method (cit). What makes the package flexible? For example, it seems that the package allows to accommodate user-defined estimation methods for the individual treatment effects. Also, what are the available models/algorithms  for estimating the other quantities? Can users specify which learners to use in the ensemble?

2. The **Algorithm** section gives a brief overview of the main steps involved in the estimation algorithm of the CRE method. Although I agree it should be a summary, some of the things stated lack context, making it unclear. Here are my comments:

- Not all elements of the equation are defined. Here is a suggestion: Line 50 "Causal Rule Ensemble relies on the Treatment Effect linear decomposition assumption, which characterizes the Conditional Average Treatment Effect (CATE) $\tau(x)$ as **the sum of** $M + 1$ distinct contributions"

- I think that a high-level overview of the algorithm in pseudo-code (as the one given in [Bargagli-Stoffi et al. (2023)](https://arxiv.org/pdf/2009.09036.pdf) ) would help readers to understand what the package does.

- line 55. "each observation is used for only one of the two steps (honest splitting)." `->` It would be useful to mention that the user can control the proportion of data used for the discovery and the inference procedures. Also, is the splitting done at random?

- What is a *fit-the-fit* procedure? This is non-standard terminology. I'd suggest giving a brief explanation and precise references for the details (e.g. section xx of paper abc).

- Similar to above. What is the "Stability Selection algorithm"?

3. Examples in the Usage section need some more context.

- Line 66. The function `generate_cre_dataset` is a function to generate a synthetic dataset. According to which model?

- Example 1. Define what `y, x, z` arguments are in `cre()` function

- Example 2. Explain what the `ite` argument expects and what is the default.

- Example 3. No need to explain all the arguments but here it would be helpful to describe what the chuck of code does and refer to the algorithm phases (e.g. discovery vs inference).

Minor
- line 54. "CRE procedure" `->` The CRE procedure

##  README

Note: I have updated this comments in view of your updates in the [https://github.com/NSAPH-Software/CRE/tree/JOSS](JOSS branch)

1. Give some context to https://github.com/NSAPH-Software/CRE/tree/main#simulations ; what are those files for? What are the simulations for?

2. What are the available options for the models? It seems that there are $3$ main models:

- a model to estimate the individual treatment effect
- a model for propensity score
- a model for outcome regression

and these are used twice, under the discovery and inference phases of the algorithm. So what models are available for each estimated quantity?  Is the same model used in both phases or one can specify different models in the two phases? Maybe a table of some sort could help to clarify this.
