### 2023-08-31 #5587 @spholmes
Thanks so much @salleuska for your help in making the paper better and I await a note from @Naeemkh to let me know when the issues mentioned have been dealt with, especially the JOSS version of the software.
