### 2023-09-08 #5587 @Naeemkh
Hi @salleuska,
Thank you for the time you allocated to review our package. All the comments and feedback were extremely valuable and helped us a lot in further improving the package and its documentation. We report here our answers to our comments point by point.

**Software Paper**
_1. The text in the Summary section seems more like a statement of need. I'd suggest revising to better highlight the package features rather than the motivations for the methodology. It is mentioned that the CRE package is a flexible implementation of the Causal Rule Ensemble method (cit). What makes the package flexible? For example, it seems that the package allows to accommodate user-defined estimation methods for the individual treatment effects. Also, what are the available models/algorithms for estimating the other quantities? Can users specify which learners to use in the ensemble?_

  - CRE package allows for multiple variants of Causal Rule Ensemble algorithm, including different internal individual average treatment effect (IATE) estimators. We have just revisited the summary accordingly.

_2. The Algorithm section gives a brief overview of the main steps involved in the estimation algorithm of the CRE method. Although I agree it should be a summary, some of the things stated lack context, making it unclear. Here are my comments:
Not all elements of the equation are defined. Here is a suggestion: Line 50 "Causal Rule Ensemble relies on the Treatment Effect linear decomposition assumption, which characterizes the Conditional Average Treatment Effect (CATE) tau(x) as the sum of  M+1 distinct contributions"_

- Good point. We expanded the description of this equation accordingly.

_I think that a high-level overview of the algorithm in pseudo-code (as the one given in [Bargagli-Stoffi et al. (2023)](https://arxiv.org/pdf/2009.09036.pdf) ) would help readers to understand what the package does._

- Thank you for the suggestion. We have just included the schematic summary of CRE already presented in the method paper.

_Line 55. "each observation is used for only one of the two steps (honest splitting)." -> It would be useful to mention that the user can control the proportion of data used for the discovery and the inference procedures. Also, is the splitting done at random?_

- We thank you for pointing this out. We have added a sentence to describe these details.

_What is a fit-the-fit procedure? This is non-standard terminology. I'd suggest giving a brief explanation and precise references for the details (e.g. section xx of paper abc)._

- With the 'fit-the-fit' procedure, we refer to the regression of the covariates based on the IATE **estimates** recently obtained. We have updated the text in the paper to reflect this. For more details about the algorithm, we suggest referring to the method paper.

_Similar to above. What is the "Stability Selection algorithm"?_

- Please refer to the citations in the paper for Meinhausen and Buhlmann 2010 and the CRE method paper for a comprehensive overview of the Stability Selection algorithm and its variations. JOSS readers are encouraged to delve into the method paper for a detailed scientific understanding.

_3. Examples in the Usage section need some more context.
Line 66. The function generate_cre_dataset is a function to generate a synthetic dataset. According to which model?_

- Good point. We added in the paper more details regarding this data generating process.

_Example 1. Define what y, x, z arguments are in cre() function._

- We have just added a description of these 3 arguments in the description of the Data Generating Process.

_Example 2. Explain what the ite argument expects and what is the default._

- We have just extended the description of Example 2, including 2 motivations to use this argument.

_Example 3. No need to explain all the arguments but here it would be helpful to describe what the chunk of code does and refer to the algorithm phases (e.g. discovery vs inference)._

- Thank you. We have enriched the description accordingly and added more references to the website.

_Line 54. "CRE procedure" -> The CRE procedure_

- Fixed. Thanks.

**README**
_Note: I have updated this comments in view of your updates in the [https://github.com/NSAPH-Software/CRE/tree/JOSS](JOSS branch)_

- Yes, that is the branch which has been updated and is intended for JOSS submission.

_Give some context to https://github.com/NSAPH-Software/CRE/tree/main#simulations ; what are those files for? What are the simulations for?_

- We added more simulation descriptions both in the main README file and the one in the functional_tests folder.

_What are the available options for the models? It seems that there are 3 models: a model to estimate the individual treatment effect, a model for propensity score and a model for outcome regression and these are used twice, under the discovery and inference phases of the algorithm. So what models are available for each estimated quantity? Is the same model used in both phases or one can specify different models in the two phases? Maybe a table of some sort could help to clarify this._

- The same IATE estimator is used for both the discovery and inference steps. The list of implemented IATE estimators is already available in the README. Each IATE estimator also requires an outcome learner and/or a propensity score learner. Both these models are simple classifiers/regressors. By default, XGBoost from the `superlearner` package is used for both these steps. We have just added these additional details in the main README file and the list of available IATE estimators in the paper.

Thank you,
The authors.
