### 2023-09-11 #5587 @salleuska
Thanks @Naeemkh. I'll go through your replies in the next few days. Can you generate the updated pdf version of the paper?

I'll leave some issues regarding the code in the JOSS branch and reference those here.
