### 2023-09-18 #5587 @salleuska
Also your answers to my points seems reasonable, but I will have to see the new version of the paper. I have one comment about this.

> The same IATE estimator is used for both the discovery and inference steps. The list of implemented IATE estimators is already available in the README. Each IATE estimator also requires an outcome learner and/or a propensity score learner. Both these models are simple classifiers/regressors. By default, XGBoost from the superlearner package is used for both these steps. We have just added these additional details in the main README file and the list of available IATE estimators in the paper.

I saw this in the Github readme, but I still have a couple of questions

1. I see for each method for the ITE estimation you linked the original paper; I would recommend to give a reference of the package you are using.
2. What are the methods available for the propensity score model and the outcome model? It is still unclear whether you are providing multiple options also for those models or you always use `SL.xgboost`
