### 2023-10-28 #5587 @salleuska
Hi @Naeemkh,

I finally was able to go through paper and software updates.

### Paper

Thanks for incorporating my suggestions, especially regarding the algorithm section. I have found some minor typos that you may want to address, but  thechecklist for the paper is complete for me.

Minor
- line 65. "( fit-the fit approach) `->` remove extra space in "( "
- line 78 "indidual treatment effect" `->` individual
- line 79 "dataset characteristic" `->` I think it should be  "characteristic**s**" ?

### Software documentation

Note: Next time you are submitting at JOSS and reviewing your code, I would recommend pointing out the branch containing the updated version in your reply. It is also good practice to include a link to the commit addressing each of the reviewer's comments; it makes the process way easier.

**General comment**

I think that the help documentation is now way clearer, but there are still some odds and ends to wrap up in terms of style. I think is important in order to make the package easily accessible to users in the R community.

In particular,

- Use quotation marks sparingly: quotation marks are used for R objects in text rendition. (see [https://developer.r-project.org/Rds.html](https://developer.r-project.org/Rds.html) )

- text that contains valid R code should be marked as such using backticks (see [https://style.tidyverse.org/documentation.html#r-code](https://style.tidyverse.org/documentation.html#r-code) for examples)
and the help for functions in base R.

References:
[https://style.tidyverse.org/documentation.html](https://style.tidyverse.org/documentation.html)
[https://developer.r-project.org/Rds.html](https://developer.r-project.org/Rds.html)

`generate_cre_dataset`

- The input parameters `->` The arguments

- I'd suggest to move into the `Details` section all the text after

"The covariates matrix is generated"

- Please remove the statement:

"In summary, this function empowers researchers and data scientists to simulate complex causal scenarios with various degrees of heterogeneity and confounding, providing valuable synthetic datasets for causal inference studies."

which is unnecessary and non-informative.

- Make sure to use backticks.s font for all the values that an argument can take. E.g. `TRUE` and also character arguments such as "lin" should be \`"lin" \` (which will be rendered in the verbatim font)

`cre` function

Thanks for clarifying the arguments' values; I think it is now clearer. As above, make sure those are rendered using backticks (ratio_dis, ite_method, etc..). This includes also **all the character values** these can take. E.g. `ite_method` values such  'aipw', 'cf', etc. This would allow one to easily spot what are the options available.
