### 2023-10-29 #5587 @Naeemkh
Thank you, @salleuska, for the informative review. We will go through your comments and address them in the final draft. Just for your information, since you mentioned that "the checklist for the paper is complete for me," it seems there is one unchecked box: "Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?" Please check that one as well. Thanks.
