### 2023-10-31 #5587 @Naeemkh
Hi @salleuska,

Thank you for taking the time to provide feedback. We have polished the paper and the documentation in accordance with your suggestions. Please let us know if there are any other issues. If so, please open a new issue for each individual task, as JOSS recommends. Adopting this approach will make it easy to track the issues and changes for each one.

Best regards,
The authors
