### 2023-11-05 NSAPH-Software/CRE#147 @salleuska
Correct roxygen documentation to reflect the type of character options for arguments
https://github.com/NSAPH-Software/CRE/blob/327487b5bdb1f4299b9cc02134ee05cd0b571ff4/R/generate_cre_dataset.R#L65-L68

I believe that options (lin, nolin, no) for the argument `confounding` are of type character in R; those should be displayed as such (using double-quotes) using verbatim font (either using `\code()` or the backticks).
E.g.

```
#' - \code{"lin"} for linear confounding
```
 and so on for the other option. Please double-check the format for the rest of the documentation.

https://github.com/openjournals/joss-reviews/issues/5587#issue-1771524251
