### 2023-11-05 NSAPH-Software/CRE#148 @salleuska
Documentation formatting
https://github.com/NSAPH-Software/CRE/blob/327487b5bdb1f4299b9cc02134ee05cd0b571ff4/R/cre.R#L1C1-L66

Options of type "character", should be displayed as such (using double-quotes) using verbatim font (either using \code() o the backticks)

I also think it could be clearer to have also the parameters of the various lists in verbatim font.

For example:

```
- \code{ratio_dis}: ...
....
- \code{ite_method}: The method for ITE (pseudo-outcome) estimation
     (default: \code{"aipw"}, options: \code{"aipw"} for Augmented Inverse
     Probability Weighting, \code{"cf"} for Causal Forest, \code{"bart"} for
     Causal Bayesian
```

Please make sure to address this also in all the other cases in the doc for the also all the other cases.

https://github.com/openjournals/joss-reviews/issues/5587#issue-1771524251
