### 2023-11-05 #5587 @salleuska
Hi @Naeemkh,

thanks for addressing the typos in the paper and referencing the commits.

Please see the issues referenced above regarding the documentation, hope now it is clearer. Again the rationale behind these requests is to allow R users to easily spot arguments and options of your functions.  I made examples in the commits, but please be sure that the format is consistent across all the documentation. Once these are addressed I'll be happy to check out the last box.
