### 2023-12-06 #5587 @Naeemkh
Thanks @spholmes, we found two other references and corrected them.

We released the package on CRAN: https://cran.r-project.org/web/packages/CRE/index.html (v0.2.5)
Also we archived the release on Zenodo: https://zenodo.org/records/10278296
DOI: [10.5281/zenodo.10278296](https://doi.org/10.5281/zenodo.10278296)

Please let us know if we need to take any other actions.
Thanks,
The authors
