### 2023-12-07 fe10df3
drop duplicated reference
