+++ paper/paper.bib
@@ -112,16 +112,6 @@ doi = {10.1214/19-BA1195},
 URL = {https://doi.org/10.1214/19-BA1195}
 }
 
-
-@article{wang2022causal,
-  title={Causal rule sets for identifying subgroups with enhanced treatment effects},
-  author={Wang, Tong and Rudin, Cynthia},
-  journal={INFORMS Journal on Computing},
-  year={2022},
-  publisher={INFORMS},
-  doi={10.1287/ijoc.2021.1143}
-}
-
 @inproceedings{nagpal2020interpretable,
   title={Interpretable subgroup discovery in treatment effect estimation with application to opioid prescribing guidelines},
   author={Nagpal, Chirag and Wei, Dennis and Vinzamuri, Bhanukiran and Shekhar, Monica and Berger, Sara E and Das, Subhro and Varshney, Kush R},
