### 2023-12-07 #5587 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
