### 2023-12-15 2bd48f8
fix typos, add affiliation
